improved responsivePicture component
